pyproject.toml
[project]
name = "rlm"
authors = [
    {name = "Alex Zhang", email = "altzhang@mit.edu"},
]
version = "0.1.0"
description = "Recursive Language Models."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.75.0",
    "google-genai>=1.56.0",
    "openai>=2.14.0",
    "opencv-python-headless>=4.13.0.90",
    "pillow>=12.1.0",
    "portkey-ai>=2.1.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "rich>=13.0.0",
    "timm>=1.0.24",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "ultralytics>=8.4.7",
]

[project.optional-dependencies]
modal = ["modal>=0.73.0", "dill>=0.3.7"]
prime = ["prime-sandboxes>=0.2.0", "dill>=0.3.7"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["rlm", "rlm.*"]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "ruff>=0.14.10",
    "ty>=0.0.7",
]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
